requirements.txt
Flask
Flask_Cors
Flask_JWT_Extended
pymongo[srv]
gunicorn
python-dotenv
cryptography